Lab3

+++ LibraryCPP/queue.h
@@ -3,7 +3,7 @@
 
 // Queue
 // Stores integer values inside
-typedef int Data;
+typedef size_t Data;
 
 struct Queue;
